Documentation for 4.0.0
It's like original but status of automated builds and status of packaging was removed. Status for discord is still there if there is internet, if not there is no text link, it's just gone. All other links are broken links if you indeed are offline. It can be nice to have in /usr/doc if you don't want to bother with md files.